requirements.txt
Flask==2.2.5
pillow==10.1.0
numpy==1.26.2
pandas==2.2.2
werkzeug==2.2.3
